implement customer methods as trait
